Moves argument parser back
