+++ P8/DataConverters/ShapeletHistogramConverter/FeatureExtraction/CMakeLists.txt
@@ -8,5 +8,5 @@ add_executable(FeatureExtraction main.cpp
         include/cxxopts/cxxopts.hpp
         include/indicators/block_progress_bar.hpp include/indicators/cursor_control.hpp
         src/FileHanding.h src/Types.h src/WindowGeneration.h src/SeriesUtils.h src/InformationGain.h src/FeatureFinding.h src/Attributes.h src/Logger.h
-        src/arguments/ArgumentParser.h)
+        src/ArgumentParser.h)
 target_include_directories(FeatureExtraction PRIVATE include)
